requirements.txt
opencv-python
matplotlib
numpy
easyocr
ultralytics
requests
Pillow
numpy
streamlit